fix(search): type search score as double
The exact-search handler now scores a hit by confidence rather than
returning a constant, emitting fractional bands (1 certain, 0.75 strong,
0.25 deep) so clients can rank an unambiguous match above fuzzy results
and hold the deep catalog tail below them.

`score` was annotated `int64`, which is correct for the fuzzy pass (a
Typesense `text_match` magnitude) but cannot represent those bands. A
generated consumer takes the annotation literally: the Kotlin client
decodes the field as `kotlin.Long`, so a fractional score is not
representable there at all.

Widen to `format: double`, which still covers the fuzzy magnitudes at the
precision JSON already limits them to. Adds a `double` helper alongside
`float`/`int64`, and corrects the `int64` doc comment, which described
itself as a double.

+++ projects/api/src/contracts/search/schema/response/searchListResponseSchema.ts
@@ -1,9 +1,9 @@
-import { int64, z } from '../../../_internal/z.ts';
+import { double, z } from '../../../_internal/z.ts';
 import { listResponseSchema } from '../../../models/index.ts';
 
 /** Zod schema for the search list response. */
 export const searchListResponseSchema = z.object({
-  score: int64(z.number().int()),
+  score: double(z.number()),
   type: z.literal('list'),
   list: listResponseSchema.nullish(),
 });
