+++ projects/api/src/contracts/search/schema/response/searchPersonResponseSchema.ts
@@ -1,9 +1,9 @@
-import { int64, z } from '../../../_internal/z.ts';
+import { double, z } from '../../../_internal/z.ts';
 import { personResponseSchema } from '../../../people/schema/response/personResponseSchema.ts';
 
 /** Zod schema for the search person response. */
 export const searchPersonResponseSchema = z.object({
-  score: int64(z.number().int()),
+  score: double(z.number()),
   type: z.literal('person'),
   person: personResponseSchema.nullish(),
 });
