+++ projects/api/src/contracts/search/schema/response/searchShowResponseSchema.ts
@@ -1,9 +1,9 @@
 import { showResponseSchema } from '../../../_internal/response/showResponseSchema.ts';
-import { int64, z } from '../../../_internal/z.ts';
+import { double, z } from '../../../_internal/z.ts';
 
 /** Zod schema for the search show response. */
 export const searchShowResponseSchema = z.object({
-  score: int64(z.number().int()),
+  score: double(z.number()),
   type: z.literal('show'),
   show: showResponseSchema.nullish(),
 });
